requirements.txt
aiofiles==0.8.0
aiogram==3.1.1
aiohttp==3.8.6
aioschedule==0.5.2
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.0.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.28.0
asyncpraw==7.7.1
asyncprawcore==2.3.0
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
environs==9.5.0
exceptiongroup==1.1.3
fake-headers==1.0.2
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
html5lib==1.1
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
magic-filter==1.0.12
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
outcome==1.2.0
packaging==23.2
pluggy==1.3.0
pydantic==2.4.2
pydantic_core==2.10.1
PySocks==1.7.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.21
tabulate==0.9.0
tenacity==8.2.3
tomli==2.0.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzlocal==5.1
update-checker==0.18.0
urllib3==2.0.6
webencodings==0.5.1
wsproto==1.2.0
yarl==1.9.2
